Retain command message for apex.js
Previously the apex command would delete the user message which invoked
it to keep the channel clean. However, not being able to see how
commands are invoked is an impediment for teaching others how commands
can be used and dissuades others from using them. Probably better to
have a noiser channel but a more utilized bot.