Refactor apex.js for better arg parsing
Made a number of improvements to apex argument parsing. Command will now
more intelligently work out the intended squad size without providing a
squad type argument. Currently you can tag yourself however and end up
in a squad twice. This is not intended use but it would be nice to
provide some logic for it in the future.