build: use standard-version for semver control
Like semantic-release, but without the automated pushing or assumption
that I'm using CI. I may move to a continuous integration system at some
point, but for the time being (deploying uisng git hooks to another
machine on the network) just bumping the version number in package.json
and generating a changelog is sufficient. Uses Conventional Changelog
under the hood.