fix(changelog): enable sending long changelogs
Sending changelog was crashing on previous version because the length of
two of the sections was over the 1024 characters allowed for a field
value. I have instead split the message into separate embeds for each
section (plus a header section) and used the description field instead,
which supports 2048 characters per embed.

Resolves #19