feat(help): upgrade help system
Structured exported command information into object literals. I will
make these class-based sometime soon, but it will be part of a larger
refactor outside of the scope of this commit which will pave the way for
moving to an eventemitter-based architecture.

Made numerous cosmetic changes to the display of help embeds.

BREAKING CHANGE: format of exported properties