fix(launcher): use node rather than nodemon
Nodemon is great for development, but it stops the app being restarted
by systemd because it captures errors and waits for changes to files,
rather than letting systemd (or whichever launch daemon) from restarting
the service. I have left the `deploy` script as the default launch
option in the service file and changed the command to `node`, and
created a separate `staging` command which uses `nodemon` instead.