chore(changelog): fix grammatical error
I needed to add tags to this branch in order to correct an issue with
standard-version creating duplicate entries in the changelog, but I
accidentally created an annotated tag, creating a link to the tag on
GitHub and breaking the changelog parser when automatically posting to
Discord. Manually correcting the changelog - the tags will be created by
standard-version going forward and thus should never be annotated.