refactor(commands): send command name with exec
Not a breaking change and no new functionality added with this commit,
but sending the name of the actual command executed allows aliases of
that command to provide different functionality. This reduces
duplication for very similar commands, e.g. a 'move message' command
could either copy the message or copy and delete the original depending
on which version of the command was used.