fix(ping): removed 'quote' alias
Removed the `quote` alias from ping because it will now be used as part
of the `move` command. This is a breaking change, but the ping command
is just for flavour and hardly used (more often it is triggered by the
bot actully being pinged), whereas the new command will be much more
useful.

BREAKING CHANGE: `quote` alias moved from `ping` command