feat(move): add `move` command
This command allows copying or moving messages between channels in a
Discord server. This is useful for keeping channels on-topic, or adding
to a best-of channel, or drawing attention to an earlier message which
users may have missed.

Currently it only works for vanilla messages, more work is required for
it to sucessfully forward rich embed posts (currently it just posts a
blank image when this is the case).