refactor(dolores): refactor commands to dolores
Significant refactor of codebase to refactor Discord-initiated commands
to a module called Dolores. Commands recognised by Discordjs now end
with a `.dolores.js` extension to allow them to be recognised and
parsed, and then added to the set of commands. Further refactoring will
convert the core functionality of these modules to separate core files
within each command folder, which will then allow different interfaces
to initiate them with minimal code duplication. For example, the `say`
command could be run from the command line if provided the necessary
guild and channel arguments, or the `apex` command could be used to
generate a list of legends for a REST API, all using the same core
function but with adapters for the different interfaces.

I chose the name Dolores because I plan to name each of these modules
after characters in the Westworld series. Dolores' role in Westworld is
to welcome and interact with guests. She is the oldest guest in the
park, just as the Discord interface is the oldest part of this app.
Convinently both Dolores and Discord also begin with D, so this is a
convinent place to start.