refactor(dolores): change `context` to `message`
I have been using the argument `context` to pass in the message object
to commands, the idea being that I could swap out message for some other
context to issue the commands from other interfaces. This isn't needed
anymore as each interface will have its own command file to deal with
it. Discordjs uses `message` for this purpose, so I'm switching back to
that to keep things clearer.