refactor(dolores): remove MessageEmbed class
Dolores#MessageEmbed was intended as a convinience method for
futureproofing, because it abstracted the Discord#MessageEmbed class
away from the discord.js library so I'd be able to make programmatic
changes to each embed based on which command required it. As it happens
I've moved to separate command files and using discord.js directly in
these files, so it is no longer needed.